11/25/19 version. Added searching functionality
Can check if we are going to win and place it there. Can also check if we can't win, if we will lose, and stop us from losing. Only looks one move ahead. Potential bug with diagNeg with consecutive checking. Also sometimes hits an infinite loop but no debugging has been done on this yet. Next step is to keep debugging, and then allow it to look ahead into future moves and pick the best option based on potential outcomes.